Add files via upload
This is an ultrasonic range finder using uart0 communication to display the distance onto a serial port on the pc using terra term and a buadrate of 57600 and turning on different led colors depending on the distance. The range we wanted for this was 0 - 100 cm anything greater than that would be out of range and displayed onto the serial port communication.